+++ arbiter-bindings/src/lib.rs
@@ -4,7 +4,3 @@ pub mod bindings;
 #[allow(clippy::all)]
 #[rustfmt::skip]
 pub mod solstat_bindings;
-
-#[allow(clippy::all)]
-#[rustfmt::skip]
-pub mod console;
